不要な分岐を削除した

+++ lib/s3_file_cache_download.rb
@@ -37,14 +37,11 @@ module S3FileCacheDownload
       S3FileCache.transaction do
         s3_file_cache = S3FileCache.find_by(s3_full_path: path, bucket_name: bucket_name)
 
-        if s3_file_cache.present?
-          if s3_file_cache.expire?
-            FileUtils.rm(s3_file_cache.place)
-            s3_file_cache.destroy
+        if s3_file_cache.present? and s3_file_cache.expire?
+          FileUtils.rm(s3_file_cache.place)
+          s3_file_cache.destroy
 
-            download_file_on_s3(path, bucket_name)
-          else
-          end
+          download_file_on_s3(path, bucket_name)
         else
           download_file_on_s3(path, bucket_name)
         end
